# komplett-server-dev/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Node-Typescript API",
    "version": "1.0.0",
    "description": "A sample API"
  },
  "servers": [
    {
      "url": "http://192.168.0.119:3000"
    },
    {
      "url": "https://mongo-kompletee.herokuapp.com"
    },
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "http://ec2-18-216-130-99.us-east-2.compute.amazonaws.com:3000"
    },
    {
      "url": "http://ec2-3-22-60-204.us-east-2.compute.amazonaws.com:3000"
    }
  ],
  "paths": {
...
    "/api/material": {
      "get": {
        "description": "Get Materials",
        "tags": [
          "material"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": 1
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": 10
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "description": "Example = Engineering"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Example = _id"
          }
        ],
        "responses": {
          "200": {
            "description": "return materials based on conditions",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MaterialSchema"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "add material",
        "tags": [
          "material"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "material created",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MaterialSchema"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "material"
        ],
        "requestBody": {
          "required": true,
          "description": "_id field is required",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialSchema"
              },
              "example": {
                "_id": "5f85845341855c2b209ba312",
                "type": "Engineering",
                "subMaterials": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MaterialSchema"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/material/withsubmaterials": {
      "post": {
        "description": "add material",
        "tags": [
          "material"
        ],
        "requestBody": {
          "required": [
            "name"
          ],
          "description": "buildingComponents and type is required",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialSchema"
              },
              "example": {
                "subMaterials": [
                  "5fa8cbacb992bf182cfec886"
                ],
                "buildingComponents": "name here",
                "type": "Engineering"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "material created",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MaterialSchema"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/material/{id}": {
      "delete": {
        "description": "delete material by id",
        "tags": [
          "material"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "content": null
          }
        }
      }
    },
    "/api/material/subMaterial": {
      "post": {
        "description": "add submaterials",
        "tags": [
          "submaterial"
        ],
        "requestBody": {
          "description": "materialId field is required",
          "required": true,
          "content": {
            "application/json": {
              "materialId": {
                "type": "string"
              },
              "schema": {
                "$ref": "#/components/schemas/InsertSubMaterial",
                "materialId": {
                  "type": "string"
                }
              }
            }
          }
        },
        "response": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InsertSubMaterial"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "update subMaterial",
        "tags": [
          "submaterial"
        ],
        "requestBody": {
          "description": "_id field is required",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubMaterialSchema"
              },
              "example": {
                "_id": "5f83da51fb2d502d6ca74b2c",
                "NS3420": "std"
              }
            }
          }
        }
      }
    },
    "/api/material/subMaterial/{materialId}": {
      "get": {
        "description": "get submaterials for the given materialId",
        "tags": [
          "submaterial"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "materialId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": {
                    "$ref": "#/components/schemas/SubMaterialSchema"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MaterialSchema": {
        "properties": {
          "crowd": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "buildingComponents": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "subMaterials": {
            "type": "array",
            "example": [],
            "items": {
              "$ref": "#/components/schemas/SubMaterialSchema"
            }
          }
        }
      },
      "SubMaterialSchema": {
        "properties": {
          "subject": {
            "type": "string"
          },
          "NS3420": {
            "type": "string"
          },
          "application": {
            "type": "string"
          },
          "crowd": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "itemPrice": {
            "type": "number"
          },
          "itemFactor": {
            "type": "number"
          },
          "inverseFactor": {
            "type": "number"
          },
          "useListPrice": {
            "type": "number"
          },
          "documents": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "quantity": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "quantityPerComponent": {
                "type": "number"
              },
              "svinn": {
                "type": "number"
              },
              "inTotal": {
                "type": "number"
              }
            }
          },
          "time": {
            "type": "object",
            "properties": {
              "minPerComponent": {
                "type": "number"
              },
              "hoursPerComponent": {
                "type": "number"
              },
              "totalTimeConsumption": {
                "type": "string"
              }
            }
          }
        }
      },
      "InsertSubMaterial": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SubMaterialSchema"
          },
          {
            "type": "object",
            "properties": {
              "materialId": {
                "type": "string"
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-token"
      }
    }
  },
...
}